rewrite config and start auth stuff
