+++ almoapi/endpoints/core/router.py
@@ -5,11 +5,11 @@ from fastapi import APIRouter, Depends, HTTPException, Request, Response
 from sse_starlette import EventSourceResponse
 
 from common import model, sampling
-from common.auth import check_admin_key, check_api_key, get_key_permission
+from auth import check_admin_key, check_api_key, get_key_permission
 from common.downloader import hf_repo_download
 from common.model import check_embeddings_container, check_model_container
 from common.networking import handle_request_error, run_with_request_disconnect
-from common.tabby_config import config
+from config.config import config
 from common.templating import PromptTemplate, get_all_templates
 from common.utils import unwrap
 from common.health import HealthManager
